Revert to FF4, update Exo
