+++ app/src/main/java/tv/danmaku/ijk/media/player/IjkMediaPlayer.java
@@ -55,6 +55,7 @@ import java.lang.ref.WeakReference;
 import java.lang.reflect.Field;
 import java.net.HttpURLConnection;
 import java.net.URL;
+import java.nio.charset.StandardCharsets;
 import java.security.InvalidParameterException;
 import java.util.ArrayList;
 import java.util.Locale;
@@ -1114,23 +1115,8 @@ public final class IjkMediaPlayer extends AbstractMediaPlayer {
                     if (msg.obj == null) {
                         player.notifyOnTimedText(null);
                     } else {
-//                        IjkTimedText text = new IjkTimedText(new Rect(0, 0, 1, 1), (String) msg.obj);
-//                        player.notifyOnTimedText(text);
-                        if (msg.arg1 == 0) {// normal
-                            IjkTimedText text = new IjkTimedText(new Rect(0, 0, 1, 1), (String) msg.obj);
-                            player.notifyOnTimedText(text);
-                        } else if (msg.arg1 == 1) { // ass
-                            IjkTimedText text = new IjkTimedText(new Rect(0, 0, 1, 1), (String) msg.obj);
-                            player.notifyOnTimedText(text);
-                        } else if (msg.arg1 == 2) { // bitmap
-                            if (msg.arg2 > 0 && msg.obj instanceof int[] && ((int[]) msg.obj).length == msg.arg2) {
-                                IjkTimedText text = new IjkTimedText((int[]) msg.obj);
-                                player.notifyOnTimedText(text);
-                            } else {
-                                IjkTimedText text = new IjkTimedText(null, "");
-                                player.notifyOnTimedText(text);
-                            }
-                        }
+                        IjkTimedText text = new IjkTimedText(new Rect(0, 0, 1, 1), (String) msg.obj);
+                        player.notifyOnTimedText(text);
                     }
                     return;
                 case MEDIA_NOP: // interface test message - ignore
