+++ app/src/main/java/tv/danmaku/ijk/media/player/IjkTimedText.java
@@ -22,23 +22,14 @@ public final class IjkTimedText {
 
     private Rect mTextBounds = null;
     private String mTextChars = null;
-    private int[] mBitmapData = null;
 
     public IjkTimedText(Rect bounds, String text) {
         mTextBounds = bounds;
         mTextChars = text;
     }
 
-    public IjkTimedText(int[] bitmap) {
-        this.mBitmapData = bitmap;
-    }
-
     public Rect getBounds() {
         return mTextBounds;
-    }
-
-    public int[] getBitmapData() {
-        return mBitmapData;
     }
 
     public String getText() {
