+++ app/src/main/java/xyz/doikki/videoplayer/aliplayer/AliMediaPlayer.java
@@ -7,6 +7,9 @@ import android.util.Log;
 import android.view.Surface;
 import android.view.SurfaceHolder;
 
+import androidx.media3.common.Player;
+import androidx.media3.common.VideoSize;
+
 import com.aliyun.player.AliPlayer;
 import com.aliyun.player.AliPlayerFactory;
 import com.aliyun.player.IPlayer;
@@ -18,8 +21,6 @@ import com.aliyun.player.nativeclass.MediaInfo;
 import com.aliyun.player.nativeclass.PlayerConfig;
 import com.aliyun.player.nativeclass.TrackInfo;
 import com.aliyun.player.source.UrlSource;
-import com.google.android.exoplayer2.Player;
-import com.google.android.exoplayer2.video.VideoSize;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -39,14 +40,171 @@ import xyz.doikki.videoplayer.player.AbstractPlayer;
 public class AliMediaPlayer extends AbstractPlayer implements Player.Listener {
 
     private final AliPlayer aliPlayer;
+    private final IPlayer.OnCompletionListener onCompletionListener = new IPlayer.OnCompletionListener() {
+        @Override
+        public void onCompletion() {
+//            notifyOnCompletion();
+            aliPlayer.stop();
+            //Log.e(TAG, "onCompletionListener onCompletion ");
+        }
+    };
+    private final IPlayer.OnVideoSizeChangedListener onVideoSizeChangedListener = new IPlayer.OnVideoSizeChangedListener() {
+        @Override
+        public void onVideoSizeChanged(int width, int height) {
+//            mVideoWidth = width;
+//            mVideoHeight = height;
+//            notifyOnVideoSizeChanged(width, height, 1, 1);
+            mPlayerEventListener.onVideoSizeChanged(width, height);
+
+            //Log.e(TAG, "onVideoSizeChangedListener " + width + " " + height);
+        }
+    };
+    private final IPlayer.OnSeekCompleteListener onSeekCompleteListener = new IPlayer.OnSeekCompleteListener() {
+        @Override
+        public void onSeekComplete() {
+            //Log.e(TAG, "onSeekCompleteListener onSeekComplete ");
+        }
+    };
+    private final IPlayer.OnSubtitleDisplayListener onSubtitleDisplayListener = new IPlayer.OnSubtitleDisplayListener() {
+        @Override
+        public void onSubtitleExtAdded(int i, String s) {
+            Log.d("derek110", "onSubtitleExtAdded: " + s);
+        }
+
+        @Override
+        public void onSubtitleShow(int i, long l, String s) {
+            Log.d("derek110", "onSubtitleShow: " + s);
+
+        }
+
+        @Override
+        public void onSubtitleHide(int i, long l) {
+            Log.d("derek110", "onSubtitleHide: ");
+        }
+
+        @Override
+        public void onSubtitleHeader(int i, String s) {
+            Log.d("derek110", "onSubtitleHeader: " + s);
+        }
+    };
+    private final IPlayer.OnErrorListener onErrorListener = new IPlayer.OnErrorListener() {
+        @Override
+        public void onError(ErrorInfo errorInfo) {
+            ErrorCode errorCode = errorInfo.getCode(); //错误码。
+            String errorMsg = errorInfo.getMsg(); //错误描述。
+            //出错后需要停止掉播放器。
+            aliPlayer.stop();
+            mPlayerEventListener.onError();
+        }
+    };
     private boolean isAliPlayerStart;
     private long currentPos;
     private int bufferPercent;
+    private final IPlayer.OnLoadingStatusListener onLoadingStatusListener = new IPlayer.OnLoadingStatusListener() {
+        @Override
+        public void onLoadingBegin() {
+            //Log.e(TAG, "onLoadingStatusListener onLoadingBegin ");
+//            notifyOnInfo(IMediaPlayer.MEDIA_INFO_BUFFERING_START, 0);
+        }
+
+        @Override
+        public void onLoadingProgress(int percent, float netSpeed) {
+            //Log.e(TAG, "onLoadingStatusListener onLoadingProgress percent = " + percent + " netSpeed = " + netSpeed);
+            bufferPercent = percent;
+        }
+
+        @Override
+        public void onLoadingEnd() {
+            //Log.e(TAG, "onLoadingStatusListener onLoadingEnd ");
+//            notifyOnInfo(IMediaPlayer.MEDIA_INFO_BUFFERING_END, 0);
+        }
+    };
     private long netSpeedLong;
+    private final IPlayer.OnInfoListener onInfoListener = new IPlayer.OnInfoListener() {
+        @Override
+        public void onInfo(InfoBean infoBean) {
+            InfoCode code = infoBean.getCode(); //信息码。
+            String msg = infoBean.getExtraMsg();//信息内容。
+            long value = infoBean.getExtraValue(); //信息值。
+            Log.d("derek110", "onInfo: " + code + " msg " + msg + "value " + value);
+            if (infoBean.getCode() == InfoCode.CurrentDownloadSpeed) {
+                //当前下载速度
+                netSpeedLong = infoBean.getExtraValue();
+                //Log.e(TAG, "sourceVideoPlayerInfo CurrentDownloadSpeed = " + netSpeedLong);
+            } else if (infoBean.getCode() == InfoCode.BufferedPosition) {
+//                //更新bufferedPosition
+//                mVideoBufferedPosition = infoBean.getExtraValue();
+            } else if (infoBean.getCode() == InfoCode.CurrentPosition) {
+                //更新currentPosition
+                currentPos = infoBean.getExtraValue();
+            }
+
+            //当前进度：InfoCode.CurrentPosition
+            //当前缓存位置：InfoCode.BufferedPosition
+        }
+    };
     private MediaInfo mAliyunMediaInfo;
     private List<TrackInfo> mAudioTrackInfoList;//音频
     private List<TrackInfo> mSubtitleTrackInfoList;//字幕
+    private final IPlayer.OnPreparedListener onPreparedListener = new IPlayer.OnPreparedListener() {
 
+        @Override
+        public void onPrepared() {
+            aliPlayer.start();
+            isAliPlayerStart = true;
+            mAliyunMediaInfo = aliPlayer.getMediaInfo();
+            mPlayerEventListener.onPrepared();
+
+            mAudioTrackInfoList = getTrackInfoListWithTrackInfoType(TrackInfo.Type.TYPE_AUDIO);
+//            mBitrateTrackInfoList = getTrackInfoListWithTrackInfoType(TrackInfo.Type.TYPE_VIDEO);
+//            mDefinitionTrackInfoList = getTrackInfoListWithTrackInfoType(TrackInfo.Type.TYPE_VOD);
+            mSubtitleTrackInfoList = getTrackInfoListWithTrackInfoType(TrackInfo.Type.TYPE_SUBTITLE);
+
+
+            for (int i = 0; i < mSubtitleTrackInfoList.size(); i++) {
+                Log.d("derek110", "mSubtitleTrackInfoList: " + mSubtitleTrackInfoList.get(i).getSubtitleLang());
+            }
+
+
+        }
+    };
+    //视频播放状态
+    private int mPlayState = IPlayer.unknow;
+    private final IPlayer.OnStateChangedListener onStateChangedListener = new IPlayer.OnStateChangedListener() {
+        @Override
+        public void onStateChanged(int i) {
+            mPlayState = i;
+            //          int idle = 0;
+            //          int initalized = 1;
+            //          int prepared = 2;
+            //          int started = 3;
+            //          int paused = 4;
+            //          int stopped = 5;
+            //          int completion = 6;
+
+            switch (i) {
+                case 0:
+                    break;
+                case 1:
+                    break;
+                case 2:
+                    mPlayerEventListener.onPrepared();
+                    break;
+                case 3:
+                    mPlayerEventListener.onInfo(AbstractPlayer.MEDIA_INFO_RENDERING_START, 0);
+                    break;
+                case 4:
+                    break;
+                case 5:
+                    break;
+                case 6:
+                    mPlayerEventListener.onCompletion();
+                    break;
+            }
+
+            //Log.e(TAG, "onStateChangedListener onStateChanged " + i);
+        }
+    };
 
     public AliMediaPlayer(Context context) {
         aliPlayer = AliPlayerFactory.createAliPlayer(context);
@@ -184,13 +342,13 @@ public class AliMediaPlayer extends AbstractPlayer implements Player.Listener {
     }
 
     @Override
-    public void setSpeed(float speed) {
-        aliPlayer.setSpeed(speed);
+    public float getSpeed() {
+        return aliPlayer.getSpeed();
     }
 
     @Override
-    public float getSpeed() {
-        return aliPlayer.getSpeed();
+    public void setSpeed(float speed) {
+        aliPlayer.setSpeed(speed);
     }
 
     @Override
@@ -206,66 +364,6 @@ public class AliMediaPlayer extends AbstractPlayer implements Player.Listener {
             mPlayerEventListener.onVideoSizeChanged(videoWidth, videoHeight);
         }
     }
-
-    private final IPlayer.OnLoadingStatusListener onLoadingStatusListener = new IPlayer.OnLoadingStatusListener() {
-        @Override
-        public void onLoadingBegin() {
-            //Log.e(TAG, "onLoadingStatusListener onLoadingBegin ");
-//            notifyOnInfo(IMediaPlayer.MEDIA_INFO_BUFFERING_START, 0);
-        }
-
-        @Override
-        public void onLoadingProgress(int percent, float netSpeed) {
-            //Log.e(TAG, "onLoadingStatusListener onLoadingProgress percent = " + percent + " netSpeed = " + netSpeed);
-            bufferPercent = percent;
-        }
-
-        @Override
-        public void onLoadingEnd() {
-            //Log.e(TAG, "onLoadingStatusListener onLoadingEnd ");
-//            notifyOnInfo(IMediaPlayer.MEDIA_INFO_BUFFERING_END, 0);
-        }
-    };
-
-    //视频播放状态
-    private int mPlayState = IPlayer.unknow;
-
-    private final IPlayer.OnStateChangedListener onStateChangedListener = new IPlayer.OnStateChangedListener() {
-        @Override
-        public void onStateChanged(int i) {
-            mPlayState = i;
-            //          int idle = 0;
-            //          int initalized = 1;
-            //          int prepared = 2;
-            //          int started = 3;
-            //          int paused = 4;
-            //          int stopped = 5;
-            //          int completion = 6;
-
-            switch (i) {
-                case 0:
-                    break;
-                case 1:
-                    break;
-                case 2:
-                    mPlayerEventListener.onPrepared();
-                    break;
-                case 3:
-                    mPlayerEventListener.onInfo(AbstractPlayer.MEDIA_INFO_RENDERING_START, 0);
-                    break;
-                case 4:
-                    break;
-                case 5:
-                    break;
-                case 6:
-                    mPlayerEventListener.onCompletion();
-                    break;
-            }
-
-            //Log.e(TAG, "onStateChangedListener onStateChanged " + i);
-        }
-    };
-
 
     /**
      * 根据TrackInfo.Type 获取对应的TrackInfo集合
@@ -308,114 +406,4 @@ public class AliMediaPlayer extends AbstractPlayer implements Player.Listener {
         }
         return trackInfoList;
     }
-
-
-    private final IPlayer.OnCompletionListener onCompletionListener = new IPlayer.OnCompletionListener() {
-        @Override
-        public void onCompletion() {
-//            notifyOnCompletion();
-            aliPlayer.stop();
-            //Log.e(TAG, "onCompletionListener onCompletion ");
-        }
-    };
-
-    private final IPlayer.OnVideoSizeChangedListener onVideoSizeChangedListener = new IPlayer.OnVideoSizeChangedListener() {
-        @Override
-        public void onVideoSizeChanged(int width, int height) {
-//            mVideoWidth = width;
-//            mVideoHeight = height;
-//            notifyOnVideoSizeChanged(width, height, 1, 1);
-            mPlayerEventListener.onVideoSizeChanged(width, height);
-
-            //Log.e(TAG, "onVideoSizeChangedListener " + width + " " + height);
-        }
-    };
-
-    private final IPlayer.OnSeekCompleteListener onSeekCompleteListener = new IPlayer.OnSeekCompleteListener() {
-        @Override
-        public void onSeekComplete() {
-            //Log.e(TAG, "onSeekCompleteListener onSeekComplete ");
-        }
-    };
-    private final IPlayer.OnInfoListener onInfoListener = new IPlayer.OnInfoListener() {
-        @Override
-        public void onInfo(InfoBean infoBean) {
-            InfoCode code = infoBean.getCode(); //信息码。
-            String msg = infoBean.getExtraMsg();//信息内容。
-            long value = infoBean.getExtraValue(); //信息值。
-            Log.d("derek110", "onInfo: " + code + " msg " + msg + "value " + value);
-            if (infoBean.getCode() == InfoCode.CurrentDownloadSpeed) {
-                //当前下载速度
-                netSpeedLong = infoBean.getExtraValue();
-                //Log.e(TAG, "sourceVideoPlayerInfo CurrentDownloadSpeed = " + netSpeedLong);
-            } else if (infoBean.getCode() == InfoCode.BufferedPosition) {
-//                //更新bufferedPosition
-//                mVideoBufferedPosition = infoBean.getExtraValue();
-            } else if (infoBean.getCode() == InfoCode.CurrentPosition) {
-                //更新currentPosition
-                currentPos = infoBean.getExtraValue();
-            }
-
-            //当前进度：InfoCode.CurrentPosition
-            //当前缓存位置：InfoCode.BufferedPosition
-        }
-    };
-
-
-    private final IPlayer.OnPreparedListener onPreparedListener = new IPlayer.OnPreparedListener() {
-
-        @Override
-        public void onPrepared() {
-            aliPlayer.start();
-            isAliPlayerStart = true;
-            mAliyunMediaInfo = aliPlayer.getMediaInfo();
-            mPlayerEventListener.onPrepared();
-
-            mAudioTrackInfoList = getTrackInfoListWithTrackInfoType(TrackInfo.Type.TYPE_AUDIO);
-//            mBitrateTrackInfoList = getTrackInfoListWithTrackInfoType(TrackInfo.Type.TYPE_VIDEO);
-//            mDefinitionTrackInfoList = getTrackInfoListWithTrackInfoType(TrackInfo.Type.TYPE_VOD);
-            mSubtitleTrackInfoList = getTrackInfoListWithTrackInfoType(TrackInfo.Type.TYPE_SUBTITLE);
-
-
-            for (int i = 0; i < mSubtitleTrackInfoList.size(); i++) {
-                Log.d("derek110", "mSubtitleTrackInfoList: " + mSubtitleTrackInfoList.get(i).getSubtitleLang());
-            }
-
-
-        }
-    };
-
-    private final IPlayer.OnSubtitleDisplayListener onSubtitleDisplayListener = new IPlayer.OnSubtitleDisplayListener() {
-        @Override
-        public void onSubtitleExtAdded(int i, String s) {
-            Log.d("derek110", "onSubtitleExtAdded: " + s);
-        }
-
-        @Override
-        public void onSubtitleShow(int i, long l, String s) {
-            Log.d("derek110", "onSubtitleShow: " + s);
-
-        }
-
-        @Override
-        public void onSubtitleHide(int i, long l) {
-            Log.d("derek110", "onSubtitleHide: ");
-        }
-
-        @Override
-        public void onSubtitleHeader(int i, String s) {
-            Log.d("derek110", "onSubtitleHeader: " + s);
-        }
-    };
-
-    private final IPlayer.OnErrorListener onErrorListener = new IPlayer.OnErrorListener() {
-        @Override
-        public void onError(ErrorInfo errorInfo) {
-            ErrorCode errorCode = errorInfo.getCode(); //错误码。
-            String errorMsg = errorInfo.getMsg(); //错误描述。
-            //出错后需要停止掉播放器。
-            aliPlayer.stop();
-            mPlayerEventListener.onError();
-        }
-    };
 }
