+++ app/src/main/java/xyz/doikki/videoplayer/exo/MyBundledHlsMediaChunkExtractor.java
@@ -1,25 +1,27 @@
 package xyz.doikki.videoplayer.exo;
 
 import androidx.annotation.VisibleForTesting;
+import androidx.media3.common.Format;
+import androidx.media3.common.util.Assertions;
+import androidx.media3.common.util.TimestampAdjuster;
+import androidx.media3.common.util.UnstableApi;
+import androidx.media3.exoplayer.hls.HlsMediaChunkExtractor;
+import androidx.media3.exoplayer.hls.WebvttExtractor;
+import androidx.media3.extractor.Extractor;
+import androidx.media3.extractor.ExtractorInput;
+import androidx.media3.extractor.ExtractorOutput;
+import androidx.media3.extractor.PositionHolder;
+import androidx.media3.extractor.mp3.Mp3Extractor;
+import androidx.media3.extractor.mp4.FragmentedMp4Extractor;
+import androidx.media3.extractor.ts.Ac3Extractor;
+import androidx.media3.extractor.ts.Ac4Extractor;
+import androidx.media3.extractor.ts.AdtsExtractor;
 
-import com.google.android.exoplayer2.Format;
-import com.google.android.exoplayer2.extractor.Extractor;
-import com.google.android.exoplayer2.extractor.ExtractorInput;
-import com.google.android.exoplayer2.extractor.ExtractorOutput;
-import com.google.android.exoplayer2.extractor.PositionHolder;
-import com.google.android.exoplayer2.extractor.mp3.Mp3Extractor;
-import com.google.android.exoplayer2.extractor.mp4.FragmentedMp4Extractor;
-import com.google.android.exoplayer2.extractor.ts.Ac3Extractor;
-import com.google.android.exoplayer2.extractor.ts.Ac4Extractor;
-import com.google.android.exoplayer2.extractor.ts.AdtsExtractor;
-import com.google.android.exoplayer2.extractor.ts.MyTsExtractor;
-import com.google.android.exoplayer2.source.hls.HlsMediaChunkExtractor;
-import com.google.android.exoplayer2.source.hls.WebvttExtractor;
-import com.google.android.exoplayer2.util.Assertions;
-import com.google.android.exoplayer2.util.TimestampAdjuster;
+import com.google.androidx.media3.exoplayer.extractor.ts.MyTsExtractor;
 
 import java.io.IOException;
 
+@UnstableApi
 /**
  * 作者：By 15968
  * 日期：On 2021/10/6
@@ -50,7 +52,7 @@ public class MyBundledHlsMediaChunkExtractor implements HlsMediaChunkExtractor {
             return this.extractor.read(extractorInput, POSITION_HOLDER) == Extractor.RESULT_CONTINUE;
         } catch (IOException e) {
             e.printStackTrace();
-            if (e instanceof com.google.android.exoplayer2.ParserException) {
+            if (e instanceof androidx.media3.common.ParserException) {
                 return false;
             }
             throw e;
